Added helper decorators
Decorators added:
require_attrs: check if required attrs is present in given instance

check_attrs: perform a check for given attr name in the instance with given check, and if fails, raise given exception. with checks as={name:(check, exception)}. checks can be passed as keyword-argument pair or an dictionary as a positional argument.

handle_exception: calls given function in a try-except statement, with except excepting 'excepted' argument and calls 'handler' with it.

require_libs: check for passed libs, whether they are a subclass of LibraryPlaceholder. If they are, then raise not implemented, else return function result.

cached: caches function return value.
if naive_cache is true, then it caches a result for any given args and kwargs pair.
if fallback_to_initial is True then you need to supply an initial_cache value. this would be used as a fallback value for dict.get method.

defaults: wraps a function, calls it and store its return value. if the return value is in supplied 'values_for_default', then return value from value_or_getter, else return the function return value.